+++ include/core/bus.h
@@ -64,7 +64,7 @@ namespace NesEmulator
         NesEmulator::Mode GetMode() const { return m_mode; }
         void SetSampleFrequency(unsigned sampleFrequency);
 
-        float GetAudioSample() { return m_apu.GetAudioSample(); }
+        double GetAudioSample() { return m_apu.GetAudioSample(); }
 
     private:
         Processor6502 m_cpu;
